perf: use uint8 types for numParticipants

+++ packages/nitro-protocol/contracts/ForceMove.sol
@@ -176,7 +176,7 @@ contract ForceMove is IForceMove {
         );
 
         _requireValidTransition(
-            fixedPart.participants.length,
+            uint8(fixedPart.participants.length),
             isFinalAB,
             variablePartAB,
             turnNumRecord + 1,
@@ -370,7 +370,7 @@ contract ForceMove is IForceMove {
         pure
         returns (bool)
     {
-        for (uint256 i = 0; i < addresses.length; i++) {
+        for (uint8 i = 0; i < addresses.length; i++) {
             if (suspect == addresses[i]) {
                 return true;
             }
@@ -395,14 +395,14 @@ contract ForceMove is IForceMove {
         Signature[] memory sigs,
         uint8[] memory whoSignedWhat // whoSignedWhat[i] is the index of the state in stateHashes that was signed by participants[i]
     ) internal pure returns (bool) {
-        uint256 nParticipants = participants.length;
-        uint256 nStates = stateHashes.length;
+        uint8 nParticipants = uint8(participants.length);
+        uint8 nStates = uint8(stateHashes.length);
 
         require(
             _acceptableWhoSignedWhat(whoSignedWhat, largestTurnNum, nParticipants, nStates),
             'Unacceptable whoSignedWhat array'
         );
-        for (uint256 i = 0; i < nParticipants; i++) {
+        for (uint8 i = 0; i < nParticipants; i++) {
             address signer = _recoverSigner(stateHashes[whoSignedWhat[i]], sigs[i]);
             if (signer != participants[i]) {
                 return false;
@@ -423,12 +423,12 @@ contract ForceMove is IForceMove {
     function _acceptableWhoSignedWhat(
         uint8[] memory whoSignedWhat,
         uint48 largestTurnNum,
-        uint256 nParticipants,
-        uint256 nStates
+        uint8 nParticipants,
+        uint8 nStates
     ) internal pure returns (bool) {
         require(whoSignedWhat.length == nParticipants, '|whoSignedWhat|!=nParticipants');
-        for (uint256 i = 0; i < nParticipants; i++) {
-            uint256 offset = (nParticipants + largestTurnNum - i) % nParticipants;
+        for (uint8 i = 0; i < nParticipants; i++) {
+            uint48 offset = (nParticipants + largestTurnNum - i) % nParticipants;
             // offset is the difference between the index of participant[i] and the index of the participant who owns the largesTurnNum state
             // the additional nParticipants in the dividend ensures offset always positive
             if (whoSignedWhat[i] + offset + 1 < nStates) {
@@ -530,7 +530,7 @@ contract ForceMove is IForceMove {
             );
             if (turnNum < largestTurnNum) {
                 _requireValidTransition(
-                    fixedPart.participants.length,
+                    uint8(fixedPart.participants.length),
                     [turnNum >= firstFinalTurnNum, turnNum + 1 >= firstFinalTurnNum],
                     [variableParts[i], variableParts[i + 1]],
                     turnNum + 1,
@@ -554,7 +554,7 @@ contract ForceMove is IForceMove {
     * @return true if the later state is a validTransition from its predecessor, false otherwise.
     */
     function _requireValidProtocolTransition(
-        uint256 nParticipants,
+        uint8 nParticipants,
         bool[2] memory isFinalAB, // [a.isFinal, b.isFinal]
         IForceMoveApp.VariablePart[2] memory ab, // [a,b]
         uint48 turnNumB
@@ -588,7 +588,7 @@ contract ForceMove is IForceMove {
     * @return true if the later state is a validTransition from its predecessor, false otherwise.
     */
     function _requireValidTransition(
-        uint256 nParticipants,
+        uint8 nParticipants,
         bool[2] memory isFinalAB, // [a.isFinal, b.isFinal]
         IForceMoveApp.VariablePart[2] memory ab, // [a,b]
         uint48 turnNumB,
